+++ R/lavaan_ind.R
@@ -6,6 +6,9 @@
 #'              the resulting `std.all` column. See "Value" section for more
 #'              details.
 #' @param fit lavaan fit object to extract fit indices from
+#' @param estimate What estimate to use, either the standardized
+#'                 estimate ("B", default), or unstandardized
+#'                 estimate ("b").
 #' @param nice_table Logical, whether to print the table as a
 #'                   `rempsyc::nice_table` as well as print the
 #'                   reference values at the bottom of the table.
@@ -40,15 +43,25 @@
 #' fit <- sem(HS.model, data=HolzingerSwineford1939)
 #' lavaan_ind(fit)
 
-lavaan_ind <- function(fit, nice_table = FALSE, underscores_to_arrows = TRUE,
-                       ...) {
-  x <- lavaan::parameterEstimates(fit, standardized = TRUE)
-  x <- x[which(x["op"] == ":="),]
-  x <- x[c("lhs", "rhs", "std.all", "pvalue", "est", "ci.lower", "ci.upper")]
-  if (isTRUE(underscores_to_arrows)) {
-    x$lhs <- gsub("_", " \u2192 ", x$lhs)
+lavaan_ind <- function(fit, estimate = "B", nice_table = FALSE,
+                       underscores_to_arrows = TRUE, ...) {
+  og.names <- c("lhs", "rhs", "pvalue", "est", "ci.lower", "ci.upper")
+  new.names <- c("Indirect Effect", "Paths", "p", "b", "CI_lower", "CI_upper")
+  if (estimate == "b") {
+    x <- lavaan::parameterEstimates(fit)
+  } else if (estimate == "B") {
+    x <- standardizedsolution(fit, level = 0.95)
+    og.names[4] <- "est.std"
+    new.names[4] <- "B"
+  } else {
+    stop("The 'estimate' argument may only be one of c('B', 'b').")
   }
-  names(x) <- c("Indirect Effect", "Paths", "B", "p", "b", "CI_lower", "CI_upper")
+  x <- x[which(x["op"] == ":="),]
+  x <- x[og.names]
+  names(x) <- new.names
+  if (isTRUE(underscores_to_arrows)) {
+    x[[1]] <- gsub("_", " \u2192 ", x[[1]])
+  }
   if (nice_table) {
     rlang::check_installed("rempsyc", reason = "for this feature.")
     x <- rempsyc::nice_table(x, ...)