+++ NEWS.md
@@ -4,7 +4,8 @@
 ## lavaanExtra 0.1.4.5
 * `lavaan_cov`: renamed r column to covariance since in some cases standardized residual variances were not real correlations.
 * New function: `lavaan_cor`, which is the same as `lavaan_cov` but only for actual correlations
-* `lavaan_reg`, `lavaan_ind`, and `lavaan_cov`: now also include the unstandardized regression coefficient and its confidence interval.
+* `lavaan_reg`, `lavaan_ind`, and `lavaan_cov`: new argument `estimate` which can be either "B" to obtain standardized estimates and corresponding p values and confidence interval, or "b" for unstandardized values.
+* `nice_lavaanPlot`: stars now appear not only for regression per default, but also for latent variables and covariances `c("regress", "latent", "covs")`.
 
 ## lavaanExtra 0.1.4.4
 * `nice_tidySEM`: corrected a bug whereas if the layout contained columns named c("IV", "M", "DV"), it would remove any extra columns (such as items), explictly ignoring part of the layout provided.